requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
idna==2.8
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
textblob==0.15.3
tweepy==3.8.0
urllib3==1.25.7
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dialogflow==0.7.2
Django==3.0.2
fbmq==2.5.0
Flask==1.1.1
google-api-core==1.15.0
google-auth==1.10.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
textblob==0.15.3
tweepy==3.8.0
urllib3==1.25.7
Werkzeug==0.16.0
